Gradle변경. lat, lng 자료형 double로 변경. 찾고자하는 장소 클릭했을 때, 해당 장소에 카메라 줌 이동 및  마커 추가.

+++ app/src/main/java/kr/mash_up/seoulmaps/data/model/PublicInfoDataSource.kt
@@ -15,6 +15,6 @@ object PublicInfoDataSource {
         publicServiceInterface = createRetrofit(PublicServiceInterface::class.java, BASE_URL)
     }
 
-    fun getToiletInfoService(lat: Float?, lng: Float?) = publicServiceInterface.getToiletService(lat, lng)
-    fun getSmokeInfoService(lat: Float?, lng: Float?) = publicServiceInterface.getSmokeService(lat, lng)
+    fun getToiletInfoService(lat: Double?, lng: Double?) = publicServiceInterface.getToiletService(lat, lng)
+    fun getSmokeInfoService(lat: Double?, lng: Double?) = publicServiceInterface.getSmokeService(lat, lng)
 }
